@@ -1,72 +1,67 @@
 {
-    "name": "notebookboost-api",
-    "version": "0.0.1",
-    "description": "API for creating and deploying jupyter notebooks on akash",
-    "author": "capturealpha",
-    "license": "TBD",
-    "main": "server.js",
-    "scripts": {
-        "start": "webpack --config webpack.dev.js --watch",
-        "build-dev": "webpack --config webpack.dev.js",
-        "build": "webpack --config webpack.prod.js",
-        "test": "jest"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/capturealpha/notebookboost"
-    },
-    "dependencies": {
-        "@akashnetwork/akashjs": "0.4.5",
-        "@chain-registry/assets": "^0.7.1",
-        "@cosmjs/crypto": "^0.28.11",
-        "@cosmjs/encoding": "^0.28.11",
-        "@cosmjs/math": "^0.28.11",
-        "@cosmjs/proto-signing": "^0.28.11",
-        "@cosmjs/stargate": "^0.28.11",
-        "@octokit/rest": "^18.12.0",
-        "@sentry/node": "^7.55.2",
-        "axios": "^0.27.2",
-        "cloudmos-shared": "file:../shared/",
-        "cors": "^2.8.5",
-        "cosmjs-types": "^0.5.0",
-        "date-fns": "^2.29.2",
-        "date-fns-tz": "^1.3.6",
-        "dotenv": "^12.0.4",
-        "express": "^4.17.1",
-        "express-async-handler": "^1.2.0",
-        "express-jwt": "^7.7.5",
-        "human-interval": "^2.0.1",
-        "js-sha256": "^0.9.0",
-        "jwks-rsa": "^2.1.4",
-        "markdown-to-txt": "^2.0.1",
-        "memory-cache": "^0.2.0",
-        "node-fetch": "^2.6.1",
-        "pg": "^8.7.3",
-        "pg-hstore": "^2.3.4",
-        "protobufjs": "^6.11.2",
-        "semver": "^7.3.8",
-        "sequelize": "^6.21.3",
-        "sequelize-typescript": "^2.1.5",
-        "stripe": "^10.14.0"
-    },
-    "devDependencies": {
-        "@types/cors": "^2.8.13",
-        "@types/express": "^4.17.13",
-        "@types/memory-cache": "^0.2.2",
-        "@types/node": "^16.6.0",
-        "@types/node-fetch": "^2.6.2",
-        "@types/pg": "^8.6.5",
-        "@types/uuid": "^8.3.1",
-        "alias-hq": "^5.1.6",
-        "dotenv-webpack": "^8.0.1",
-        "nodemon": "^2.0.7",
-        "nodemon-webpack-plugin": "^4.5.2",
-        "supertest": "^6.1.5",
-        "ts-jest": "^27.0.4",
-        "ts-loader": "^9.2.5",
-        "typescript": "^4.3.5",
-        "webpack": "^5.33.2",
-        "webpack-cli": "^4.7.2",
-        "webpack-node-externals": "^3.0.0"
+  "name": "launch-api",
+  "version": "0.0.1",
+  "description": "API for creating and deploying jupyter notebooks on akash",
+  "author": "capturealpha",
+  "license": "TBD",
+  "main": "server.js",
+  "predeploy": "npm version patch",
+  "scripts": {
+    "start": "webpack --config webpack.dev.js --watch",
+    "build:dev": "webpack --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "lint": "eslint --cache \"src/**/*.{js,jsx,ts,tsx}\"",
+    "lint:fix": "eslint --cache \"src/**/*.{js,jsx,ts,tsx}\" --fix",
+    "depcheck": "depcheck --ignores=typescript,ts-jest,ts-loader,webpack,webpack-cli,webpack-node-externals,alias-hq,eslint,eslint-plugin-import,eslint-plugin-node,eslint-plugin-promise,eslint-plugin-standard,eslint-plugin-jest",
+    "prepare": "cd ../ && husky install api/.husky",
+    "prettier:check": "prettier --list-different \"**/*.{js,jsx,ts,tsx,json,md}\"",
+    "prettier:fix": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
+    "update-deps": "npx npm-check-updates -u"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
     }
+  },
+  "lint-staged": {
+    "*.{ts,tsx}": "npm run lint:fix"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/capturealpha/notebookboost"
+  },
+  "dependencies": {
+    "@akashnetwork/akashjs": "0.4.9",
+    "cors": "^2.8.5",
+    "express": "^4.18.2",
+    "express-async-handler": "^1.2.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^14.0.1",
+    "ts-loader": "^9.4.4"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.13",
+    "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.4",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "eslint-config-standard-with-typescript": "^37.0.0",
+    "alias-hq": "^6.2.2",
+    "depcheck": "^1.4.5",
+    "dotenv-webpack": "^8.0.1",
+    "eslint": "^8.47.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "nodemon": "^3.0.1",
+    "nodemon-webpack-plugin": "^4.8.1",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.1.6",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4",
+    "webpack-node-externals": "^3.0.0"
+  }
 }
