@@ -1,29 +1,33 @@
 import express from "express";
 import cors from "cors";
-import DeploymentRouter from "./routers/deployment";
+// import DeploymentRouter from './routers/deployment-router'
 
 const app = express();
 app.use(
   cors({
-    origin: process.env.NOTEBOOKBOOST_CORS_WEBSITE_URLS?.split(",") || ["http://localhost:3000", "http://localhost:3001"],
+    origin: process.env.LAUNCH_CORS_WEBSITE_URLS?.split(",") != null || [
+      "http://localhost:3000",
+      "http://localhost:3001"
+    ],
     optionsSuccessStatus: 200
   })
 );
 
 const { PORT = 3080 } = process.env;
-app.use("/deployment", await DeploymentRouter());
-console.log(process.env.API_VERSION, process.env.RPC_ENDPOINT);
+// app.use('/deployment', DeploymentRouter(null))
+// console.log(process.env.API_VERSION, process.env.RPC_ENDPOINT)
 
-async function initApp() {
+async function initApp(): Promise<unknown> {
   try {
-    app.listen(PORT, () => {
+    return app.listen(PORT, () => {
       console.log(`server started at http://localhost: ${PORT}`);
     });
   } catch (err) {
     console.error("Error while initializing app", err);
+    return await Promise.reject(err);
   }
 }
 
-initApp();
+await initApp();
 
 export default app;
